timer fixed commit
