+++ src/headers/sorters/Sorter.h
@@ -10,7 +10,7 @@
 #include "BinaryInsertionSort.h"
 #include "QuickSort.h"
 #include "HeapSort.h"
-#include "utils/Helper.h"
+#include "utils/Timer.h"
 
 enum class SortType {
     InsertionSort,
@@ -23,33 +23,33 @@ template<typename Type>
 class Sorter {
 public:
     static void sort(std::vector<Type> &vec, SortType sortType) {
-        Helper helper;
+        Timer timer;
         switch (sortType) {
             case SortType::InsertionSort:
 
-                helper.timer_start();
+                timer.timer_start();
                 InsertionSort::insertionSort(vec);
-                std::cout << "Insertion Sort: czas sortowania " << vec.size() <<  "elementow : " << std::to_string(
-                    helper.timer_stop()) + "\n";
+                std::cout << "Insertion Sort: czas sortowania " << vec.size() <<  " elementow: " << std::to_string(
+                    timer.timer_stop()) + "\n";
 
                 break;
             case SortType::BinaryInsertionSort:
 
-                helper.timer_start();
+                timer.timer_start();
                 BinaryInsertionSort::binaryInsertionSort(vec);
                 std::cout << "Binary Insertion Sort: czas sortowania " << vec.size() << "elementow : " << std::to_string(
-                    helper.timer_stop()) + "\n";
+                    timer.timer_stop()) + "\n";
             case SortType::QuickSort:
-                helper.timer_start();
+                timer.timer_start();
                 QuickSort::quicksort(vec);
                 std::cout << "Quick Sort: czas sortowania " << vec.size() << "elementow : " << std::to_string(
-                    helper.timer_stop()) + "\n";
+                    timer.timer_stop()) + "\n";
                 break;
             case SortType::HeapSort:
-                helper.timer_start();
+                timer.timer_start();
                 HeapSort::heapSort(vec);
                 std::cout << "Heap Sort: czas sortowania " << vec.size() << "elementow : " << std::to_string(
-                    helper.timer_stop()) + "\n";
+                    timer.timer_stop()) + "\n";
             default:
                 return;
         }
